test(frontend): rename Home mocks and update imports
